refactor: 지도에 상세 정보 제한을 보여주는 패널을 컴포넌트로 분리
